requirements.txt
faiss-cpu
sentence-transformers
gradio
mlflow
plotly
pandas
